Rearranged repository to make it simpler to find files under their specific functions
Renamed deposit functions and folders as obsolete_deposits as "deposits" no longer in download stats es index. It might be added back in later using the fbi index.

Removed fields which are no longer present in the new index. Renamed fields which have had name changes.
