+++ access_stats/settings.py
@@ -14,7 +14,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'downloads',
-    'deposits',
     'cookielaw',
     'bootstrap_datepicker_plus',
 ]
