+++ common/json_maker_factory.py
@@ -1,12 +1,9 @@
 from common.json.methods_json import MethodsJson
 from common.json.timeline_json import TimelineJson
-from common.json.deposits_timeline_json import DepositsTimelineJson
 from common.json.dataset_json import DatasetJson
-from common.json.deposits_dataset_json import DepositsDatasetJson
 from common.json.user_json import UserJson
 from common.json.users_json import UsersJson
 from common.json.trace_json import TraceJson
-from common.json.deposits_trace_json import DepositsTraceJson
 
 class JsonMakerFactory:
     def __init__(self, deposits = False):
@@ -17,18 +14,12 @@ class JsonMakerFactory:
             return MethodsJson(filters, analysis_method)
         if analysis_method == "timeline" and not self.deposits:
             return TimelineJson(filters, analysis_method)
-        if analysis_method == "timeline" and self.deposits:
-            return DepositsTimelineJson(filters, analysis_method)
         if analysis_method == "dataset" and not self.deposits:
             return DatasetJson(filters, analysis_method)
-        if analysis_method == "dataset" and self.deposits:
-            return DepositsDatasetJson(filters, analysis_method)
         if analysis_method == "user":
             return UserJson(filters, analysis_method)
         if analysis_method == "users":
             return UsersJson(filters, analysis_method)
         if analysis_method == "trace" and not self.deposits:
             return TraceJson(filters, analysis_method)
-        if analysis_method == "trace" and self.deposits:
-            return DepositsTraceJson(filters, analysis_method)
         raise RuntimeError("Invalid analysis method")
