+++ common/query_builder_factory.py
@@ -1,12 +1,10 @@
 from common.es_queries.methods_query import MethodsQuery
 from common.es_queries.timeline_query import TimelineQuery
-from common.es_queries.deposits_timeline_query import DepositsTimelineQuery
 from common.es_queries.dataset_query import DatasetQuery
-from common.es_queries.deposits_dataset_query import DepositsDatasetQuery
 from common.es_queries.user_query import UserQuery
 from common.es_queries.users_query import UsersQuery
 from common.es_queries.trace_query import TraceQuery
-from common.es_queries.deposits_trace_query import DepositsTraceQuery
+from common.obsolete_deposits.deposits_trace_query import DepositsTraceQuery
 
 class QueryBuilderFactory:
     def __init__(self, deposits = False):
@@ -17,12 +15,8 @@ class QueryBuilderFactory:
             return MethodsQuery(filters, analysis_method)
         if analysis_method == "timeline" and not self.deposits:
             return TimelineQuery(filters, analysis_method)
-        if analysis_method == "timeline" and self.deposits:
-            return DepositsTimelineQuery(filters, analysis_method)
         if analysis_method == "dataset" and not self.deposits:
             return DatasetQuery(filters, analysis_method, after_key)
-        if analysis_method == "dataset" and self.deposits:
-            return DepositsDatasetQuery(filters, analysis_method, after_key)
         if analysis_method == "user":
             return UserQuery(filters, analysis_method)
         if analysis_method == "users":
