+++ downloads/static/downloads/js/user.js
@@ -11,49 +11,17 @@ $.get(
 function renderUserPage(data)
 {
     var labelsDict = {
-        field: [],
         country: [],
-        instituteType: [],
-        odaCountry: [],
-        area: []
     }
 
     var dataDict = {
-        field: [],
         country: [],
-        instituteType: [],
-        odaCountry: [],
-        area: []
-    }
-
-    for (var field in data.results.group_by_field)
-    {
-        labelsDict.field.push(field);
-        dataDict.field.push(data.results.group_by_field[field]);
     }
 
     for (var country in data.results.group_by_country)
     {
         labelsDict.country.push(country);
         dataDict.country.push(data.results.group_by_country[country]);
-    }
-
-    for (var instituteType in data.results.group_by_institute_type)
-    {
-        labelsDict.instituteType.push(instituteType);
-        dataDict.instituteType.push(data.results.group_by_institute_type[instituteType]);
-    }
-
-    for (var odaCountry in data.results.group_by_oda_type)
-    {
-        labelsDict.odaCountry.push(odaCountry);
-        dataDict.odaCountry.push(data.results.group_by_oda_type[odaCountry]);
-    }
-
-    for (var area in data.results.group_by_area)
-    {
-        labelsDict.area.push(area);
-        dataDict.area.push(data.results.group_by_area[area]);
     }
 
     var activeTab = null;
@@ -78,7 +46,7 @@ function renderUserPage(data)
     }
     userChart = updateUserChart(userChart, activeTab, dataDict, labelsDict);
 
-    userTabs = ["userTabField", "userTabCountry", "userTabInstituteType", "userTabOdaCountry", "userTabArea"]
+    userTabs = ["userTabCountry"]
     $('a[data-toggle="tab-sub"]').on("shown.bs.tab", function (e) {
         if (userTabs.includes(e.target.id))
         {
@@ -98,31 +66,12 @@ function updateUserChart(chart, activeTab, dataDict, labelsDict)
 
 function makeUserChart(activeTab, dataDict, labelsDict)
 {
-    if (activeTab == "userTabField" || !activeTab)
-    {
-        var activeLabels = labelsDict.field;
-        var activeData = dataDict.field;
-    }
-    else if (activeTab == "userTabCountry")
+    if (activeTab == "userTabCountry")
     {
         var activeLabels = labelsDict.country;
         var activeData = dataDict.country;
     }
-    else if (activeTab == "userTabInstituteType")
-    {
-        var activeLabels = labelsDict.instituteType;
-        var activeData = dataDict.instituteType;
-    }
-    else if (activeTab == "userTabOdaCountry")
-    {
-        var activeLabels = labelsDict.odaCountry;
-        var activeData = dataDict.odaCountry;
-    }
-    else if (activeTab == "userTabArea")
-    {
-        var activeLabels = labelsDict.area;
-        var activeData = dataDict.area;
-    }
+
     var html = Mustache.render(templates.canvas, {id:"userChart"})
     $("#userChartBox").html(html);
     var userChartElement = $("#userChart");
