+++ downloads/templates/downloads/index.html
@@ -228,21 +228,7 @@
                     <div class="d-flex flex-row mt-2">
                       <ul class="nav nav-tabs nav-tabs--vertical nav-tabs--right" id="user-tab">
                         <li class="nav-item">
-                          <a class="nav-link active" data-toggle="tab-sub" href="#userTabField" id="userTabField">Field</a>
-                        </li>
-                        <li class="nav-item">
                           <a class="nav-link" data-toggle="tab-sub" href="#userTabCountry" id="userTabCountry">Country</a>
-                        </li>
-                        <li class="nav-item">
-                          <a class="nav-link" data-toggle="tab-sub" href="#userTabInstituteType" id="userTabInstituteType">Institute Type</a>
-                        </li>
-                        {% if not public %}
-                          <li class="nav-item">
-                            <a class="nav-link" data-toggle="tab-sub" href="#userTabOdaCountry" id="userTabOdaCountry">ODA Status</a>
-                          </li>
-                        {% endif %}
-                        <li class="nav-item">
-                          <a class="nav-link" data-toggle="tab-sub" href="#userTabArea" id="userTabArea">Area</a>
                         </li>
                       </ul>
                     </div>
