+++ obsolete_deposits/static/deposits/css/deposits.css
@@ -1,0 +1,37 @@
+/* Code to get vertical tabs for second layer filtering */
+
+html {
+  overflow-y: scroll; 
+}
+
+.nav-tabs--vertical {
+  border-bottom: none;
+  border-left: 1px solid #ddd;
+  display: flex;
+  flex-flow: column nowrap;
+}
+.nav-tabs--right {
+  margin: 0 15px;
+}
+.nav-tabs--right .nav-item + .nav-item {
+  margin-top: .25rem;
+}
+.nav-tabs--right .nav-link {
+  transition: border-color .125s ease-in;
+  white-space: nowrap;
+}
+.nav-tabs--right .nav-link:hover {
+  background-color: #f7f7f7;
+  border-color: transparent;
+}
+.nav-tabs--right .nav-link.active {
+  border-bottom-color: #ddd;
+  border-left-color: #fff;
+  border-bottom-right-radius: 0.25rem;
+  border-top-left-radius: 0;
+  margin-left: -1px;
+}
+.nav-tabs--right .nav-link.active:hover {
+  background-color: #fff;
+  border-color: #0275d8 #0275d8 #0275d8 #fff;
+}
